fix the dataset download issue of sentiment analysis
